LayoutTests/imported/w3c/web-platform-tests/css/css-values/minmax-angle-serialize-expected.txt


PASS e.style['transform'] = "rotate(min(1deg))" should set the property value 
FAIL e.style['transform'] = "rotate(min(1rad))" should set the property value assert_equals: serialization should be canonical expected "rotate(min(1rad))" but got "rotate(min(57.29577951308232deg))"
FAIL e.style['transform'] = "rotate(min(1turn))" should set the property value assert_equals: serialization should be canonical expected "rotate(min(1turn))" but got "rotate(min(360deg))"
FAIL e.style['transform'] = "rotate(min(1grad))" should set the property value assert_equals: serialization should be canonical expected "rotate(min(1grad))" but got "rotate(min(0.9deg))"
PASS e.style['transform'] = "rotate(max(1deg))" should set the property value 
FAIL e.style['transform'] = "rotate(max(1rad))" should set the property value assert_equals: serialization should be canonical expected "rotate(max(1rad))" but got "rotate(max(57.29577951308232deg))"
FAIL e.style['transform'] = "rotate(max(1turn))" should set the property value assert_equals: serialization should be canonical expected "rotate(max(1turn))" but got "rotate(max(360deg))"
FAIL e.style['transform'] = "rotate(max(1grad))" should set the property value assert_equals: serialization should be canonical expected "rotate(max(1grad))" but got "rotate(max(0.9deg))"
FAIL e.style['transform'] = "rotate(min(1deg, 2deg, 3deg))" should set the property value assert_equals: serialization should be canonical expected "rotate(min(1deg, 2deg, 3deg))" but got "rotate(min(1deg))"
FAIL e.style['transform'] = "rotate(min(3deg, 2deg, 1deg))" should set the property value assert_equals: serialization should be canonical expected "rotate(min(3deg, 2deg, 1deg))" but got "rotate(min(1deg))"
FAIL e.style['transform'] = "rotate(min(90deg, 1.57rad, 0.25turn))" should set the property value assert_equals: serialization should be canonical expected "rotate(min(90deg, 1.57rad, 0.25turn))" but got "rotate(min(89.95437383553924deg))"
FAIL e.style['transform'] = "rotate(min(0.25turn, 1.57rad, 90deg))" should set the property value assert_equals: serialization should be canonical expected "rotate(min(0.25turn, 1.57rad, 90deg))" but got "rotate(min(89.95437383553924deg))"
FAIL e.style['transform'] = "rotate(max(1deg, 2deg, 3deg))" should set the property value assert_equals: serialization should be canonical expected "rotate(max(1deg, 2deg, 3deg))" but got "rotate(max(3deg))"
FAIL e.style['transform'] = "rotate(max(3deg, 2deg, 1deg))" should set the property value assert_equals: serialization should be canonical expected "rotate(max(3deg, 2deg, 1deg))" but got "rotate(max(3deg))"
FAIL e.style['transform'] = "rotate(max(90deg, 1.57rad, 0.25turn))" should set the property value assert_equals: serialization should be canonical expected "rotate(max(90deg, 1.57rad, 0.25turn))" but got "rotate(max(90deg))"
FAIL e.style['transform'] = "rotate(max(0.25turn, 1.57rad, 90deg))" should set the property value assert_equals: serialization should be canonical expected "rotate(max(0.25turn, 1.57rad, 90deg))" but got "rotate(max(90deg))"
FAIL e.style['transform'] = "rotate(calc(min(1deg) + min(2deg)))" should set the property value assert_equals: serialization should be canonical expected "rotate(calc(min(1deg) + min(2deg)))" but got "rotate(calc(3deg))"
FAIL e.style['transform'] = "rotate(calc(max(1deg) + max(2deg)))" should set the property value assert_equals: serialization should be canonical expected "rotate(calc(max(1deg) + max(2deg)))" but got "rotate(calc(3deg))"
FAIL e.style['transform'] = "rotate(calc(1rad + min(1deg)))" should set the property value assert_equals: serialization should be canonical expected "rotate(calc(1rad + min(1deg)))" but got "rotate(calc(58.29577951308232deg))"
FAIL e.style['transform'] = "rotate(calc(min(1deg) + 1rad))" should set the property value assert_equals: serialization should be canonical expected "rotate(calc(1rad + min(1deg)))" but got "rotate(calc(58.29577951308232deg))"
FAIL e.style['transform'] = "rotate(calc(1rad + max(1deg)))" should set the property value assert_equals: serialization should be canonical expected "rotate(calc(1rad + max(1deg)))" but got "rotate(calc(58.29577951308232deg))"
FAIL e.style['transform'] = "rotate(calc(max(1deg) + 1rad))" should set the property value assert_equals: serialization should be canonical expected "rotate(calc(1rad + max(1deg)))" but got "rotate(calc(58.29577951308232deg))"
